requirements.txt
PySide6>=6.6.0
PyMuPDF>=1.24.0
pyHanko[pkcs11,image-support]>=0.25.0
cryptography>=42.0.0
Pillow>=10.0.0
